Make the text size a 'float' instead of 'uint8_t'.

+++ Adafruit_GFX.h
@@ -105,17 +105,17 @@ public:
   void drawRGBBitmap(int16_t x, int16_t y, uint16_t *bitmap, uint8_t *mask,
                      int16_t w, int16_t h);
   void drawChar(int16_t x, int16_t y, unsigned char c, uint16_t color,
-                uint16_t bg, uint8_t size);
+                uint16_t bg, float size);
   void drawChar(int16_t x, int16_t y, unsigned char c, uint16_t color,
-                uint16_t bg, uint8_t size_x, uint8_t size_y);
+                uint16_t bg, float size_x, float size_y);
   void getTextBounds(const char *string, int16_t x, int16_t y, int16_t *x1,
                      int16_t *y1, uint16_t *w, uint16_t *h);
   void getTextBounds(const __FlashStringHelper *s, int16_t x, int16_t y,
                      int16_t *x1, int16_t *y1, uint16_t *w, uint16_t *h);
   void getTextBounds(const String &str, int16_t x, int16_t y, int16_t *x1,
                      int16_t *y1, uint16_t *w, uint16_t *h);
-  void setTextSize(uint8_t s);
-  void setTextSize(uint8_t sx, uint8_t sy);
+  void setTextSize(float s);
+  void setTextSize(float sx, float sy);
   void setFont(const GFXfont *f = NULL);
 
   /**********************************************************************/
@@ -237,8 +237,8 @@ protected:
   int16_t cursor_y;     ///< y location to start print()ing text
   uint16_t textcolor;   ///< 16-bit background color for print()
   uint16_t textbgcolor; ///< 16-bit text color for print()
-  uint8_t textsize_x;   ///< Desired magnification in X-axis of text to print()
-  uint8_t textsize_y;   ///< Desired magnification in Y-axis of text to print()
+  float textsize_x;   ///< Desired magnification in X-axis of text to print()
+  float textsize_y;   ///< Desired magnification in Y-axis of text to print()
   uint8_t rotation;     ///< Display rotation (0 thru 3)
   bool wrap;            ///< If set, 'wrap' text at right edge of display
   bool _cp437;          ///< If set, use correct CP437 charset (default is off)
